+++ changes.md
@@ -1,5 +1,10 @@
 ## Main wip
 
+## 0.3.1
+* Added tile bolt_panel
+* Added boltPanel example
+* Documentation for bolt_panel and glyph
+
 ## 0.3.0
 * Upped cadqueryhelper version 0.2.0
   * "Base" class refactoring